:white_check_mark: Back make voter (#41)
* :art: add userVoter and function isGranted on userController

* :art: add ServiceVoter and function isGranted on ServiceController

* :art: add route to display all services by user

* :art: add CommentVoter and function isGranted on CommentController

* add new payload, ans security for service-category

+++ oreijin-backend/src/Controller/ServiceController.php
@@ -4,6 +4,8 @@ namespace App\Controller;
 
 use App\Entity\Service;
 use App\Manager\ServiceManager;
+use App\Security\Voter\ServiceVoter;
+use App\Security\Voter\UserVoter;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -21,22 +23,6 @@ class ServiceController extends AbstractController
 
     /**
      * @Route(
-     *      "/api/services",
-     *      name="api_service_browse",  
-     *      methods={"GET"}
-     * )
-     */
-    public function browse(): Response
-    {
-        $services = $this->serviceManager->browse();
-
-        $services = $this->serviceManager->serialize($services, ['groups' => 'services-list']);
-
-        return new Response($services);
-    }
-
-        /**
-     * @Route(
      *      "/api/services/home",
      *      name="api_service_browse_home",  
      *      methods={"GET"}
@@ -45,6 +31,38 @@ class ServiceController extends AbstractController
     public function home(): Response
     {
         $services = $this->serviceManager->home();
+
+        $services = $this->serviceManager->serialize($services, ['groups' => 'services-list']);
+
+        return new Response($services);
+    }
+
+    /**
+     * @Route(
+     *      "/api/services/user/{userId}",
+     *      name="api_service_browse_user",  
+     *      methods={"GET"}, 
+     *      requirements={"id"="\d+"}
+     * )
+     */
+    public function allServicesByUser($userId): Response
+    {
+        $services = $this->serviceManager->servicesByUser($userId);
+        $services = $this->serviceManager->serialize($services, ['groups' => 'services-list']);
+
+        return new Response($services);
+    }
+
+    /**
+     * @Route(
+     *      "/api/services",
+     *      name="api_service_browse",  
+     *      methods={"GET"}
+     * )
+     */
+    public function browse(): Response
+    {
+        $services = $this->serviceManager->browse();
 
         $services = $this->serviceManager->serialize($services, ['groups' => 'services-list']);
 
@@ -76,9 +94,10 @@ class ServiceController extends AbstractController
     public function add(Request $request): Response
     {
         $data = $request->getContent();
+
         $service = $this->serviceManager->create($data);
         $service = $this->serviceManager->serialize($service, ['groups' => 'service-add']);
-        
+
         return new Response($service, Response::HTTP_CREATED);
     }
 
@@ -92,6 +111,8 @@ class ServiceController extends AbstractController
      */
     public function edit(Request $request, Service $service): Response
     {
+        $this->denyAccessUnlessGranted(UserVoter::SAME_USER, $service);
+
         $data = $request->getContent();
         $service = $this->serviceManager->update($service, $data);
         $service = $this->serviceManager->serialize($service, ['groups' => 'service-edit']);
@@ -112,6 +133,8 @@ class ServiceController extends AbstractController
      */
     public function delete(Service $service): JsonResponse
     {
+        $this->denyAccessUnlessGranted(UserVoter::SAME_USER, $service);
+
         $this->serviceManager->delete($service);
 
         return new JsonResponse(
